+++ lib/services/location_service.dart
@@ -1,0 +1,26 @@
+import 'package:geolocator/geolocator.dart';
+
+class LocationService {
+  static Future<Position?> determinePosition() async {
+    bool serviceEnabled = await Geolocator.isLocationServiceEnabled();
+    if (!serviceEnabled) return null;
+    LocationPermission permission = await Geolocator.checkPermission();
+    if (permission == LocationPermission.denied) {
+      permission = await Geolocator.requestPermission();
+      if (permission == LocationPermission.denied) return null;
+    }
+    if (permission == LocationPermission.deniedForever) return null;
+    try {
+      return await Geolocator.getCurrentPosition(
+        desiredAccuracy: LocationAccuracy.high,
+      );
+    } catch (e) {
+      return null;
+    }
+  }
+
+  static Future<String> getAddressFromPosition(Position position) async {
+    // In a real app, you might use reverse-geocoding here.
+    return 'Lat: ${position.latitude}, Lon: ${position.longitude}';
+  }
+}
